Improve usability of sharing wizard
Having an ellipsis in the "Create..." button's text gives users the assumption
that a dialog will come up before the real action is triggered hence the ellipsis
is removed.

In order to improve consistency the label for the ".git" postfix is enabled together
with the edit field for the repository to be created.

Bug: 292592
Change-Id: Ie920b3b4dc83ce35a005fe3b4bcf875d33f57499
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>