Add a "Git Repositories View"
This is a very first implementation of a "Git Repositories View"
(analogous to e.g. CVS Repositories View). It allows addition
and removal of local Git Repositories, cloning of Git
Repositories as well as checking-out of branches.

Project import can be done using the standard "Import Existing Projects"
wizard.

We should use bug 301168 to discuss this.

Bug: 301168 (also: 281394, 302742, 303403)

Change-Id: Icd0a7171f67c3180dc4a4a4246b490d6d4536845
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>