Dispose of any instantiated images when the label provider is disposed (bug 309601).
ImageDescriptors creates new SWT images every time
createImage() is invoked. Images need to be explicitly disposed
when they are no longer needed or the application will eventually
run out of resources. A map is now kept of all the images that
have been instantiated so clean-up can be performed when the
label provider is disposed.

Change-Id: Ie2530aac18ac6c70011ec3bba15278fc253599c2
Signed-off-by: Chris Aniszczyk <caniszczyk@gmail.com>