Fix leaks of SWT Color instances
FetchResultTable and PushResultTable was originally using JFace's
ColorRegistry for allocating its Color instances. The ColorRegistry
will dispose of its Colors when its owning Display is disposed. In
the case of the Eclipse workbench, the Display will only ever be
disposed when Eclipse itself shuts down.

Since the tables create new registries every time they are created,
a Color instance is leaked for every request that is made to the
registry. The code has been corrected to simply instantiate new,
local Color instances directly which are now disposed when the
table widget is disposed.

Change-Id: Ida8ffcde63a21fe92321643740b8836510a56ea8
Signed-off-By: Chris Aniszczyk <caniszczyk@gmail.com>