Use jgit p2 repository to resolve jgit dependencies
EGit Tycho builds on build.eclipse.org frequently hit corrupted artifacts
which leads to broken builds. Cleaning up these corrupted files is tedious
since it requires file system access on the build server. Hence we want to
switch to use job-local m2 repositories. This requires that build artifacts
are shared between the jgit and egit build jobs via p2. Therefore we switch
from pom-first to p2 build dependencies between the egit and the jgit build.

For local builds the egit build assumes that jgit and egit are located
under a common parent folder. If this is not the case the URL of the jgit
update site has to be injected via system property "jgit-site".

- non-default local build:
$ mvn clean install -Djgit-site=file:/path/to/org.eclipse.jgit.updatesite/target/site

- hudson build on build.eclipse.org
$ mvn clean install -Djgit-site=https://build.eclipse.org/hudson/job/jgit/
              lastSuccessfulBuild/artifact/org.eclipse.jgit.packaging/
              org.eclipse.jgit.updatesite/target/site/

Change-Id: I43ee21f23d474a6579df187869c821fc3960583f
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>