Relax constructor parameters of EGit operations
Certain EGit operations takes a Collection<IResource> as a
parameter in its constructor. The problem with this type is that
one cannot pass in a Collection<IFile> or a Collection<IFolder>.
Instead, you are forced to alter your existing collection into one
for IResources. The fix is to alter these constructors so that they
take a Collection<? extends IResource> instead.

Change-Id: I93acf875107596b7eb08d76aaf69f13ed0cab524