Ignore secondary mouse clicks in the 'History' view's hyperlinks
Only the primary mouse click should be acknowledged as a request
to jump to another commit when browsing commits in the 'History'
view. At the moment, users that want to select the commit id and
copy the id to the clipboard through the context menu might
actually end up jumping to the selected commit as the secondary
mouse click is processed as a hyperlink request.

Change-Id: I8351cbda039419160e6731e307fa7243a3be90d4