Improve error validation in the patch wizard
Currently, no error messages are presented to the user for invalid
input into the wizard. This is very unhelpful as the user may not
know what is wrong with their input. Switching to the clipboard
mode may also not clear the error message due to an error in the
validation logic. To also help prevent the user from entering data
that may not end up being processed, controls are now enabled and
disabled based on which radio button has been selected.

Change-Id: I50998266c213deabec4a01a44024d95015a56b55