Fix compare error by checking if blob is contained in commit
Opening compare editor on a deleted file in the file list 
of the History View caused a message to be logged. The
reason was that the blob was not found in the commit.

This patch makes use of code that was already available
for other compare actions. To allow for reuse, a new 
class CompareUtils is introduced.

Bug: 312852
Signed-off-by: Stefan Lay <stefan.lay@sap.com>