Allow Git repository to be pre-filled from clipboard
If the user has copied a clipboard already, then we can access that
from the SWT clipboard. If it looks like a transport we know about
(as reported by Transport.canHandleProtocol) then we pre-fill the
connection dialog with that as the URI.

Since we need to fill in the user/port combinations, refactor out
the code which does this on keypress and instead call that function,
but only once, and only after all the fields have been created.

Bug: 315589
Change-Id: I2fd6f0d68b4908f938eb5b2df8cd4f99e674e74d
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>