Refactor "previously used values" content assist
Currently, there are two places where we have implemented
a content assist with previously used values: in the commit dialog
for the author and committer fields, and in the RepositorySelection
page for the URI field. The logic should be put into one place and
be documented properly to ensure consistency, supportability,
and re-usability.

Bug: 315186
Change-Id: I21f8c992a3fbb3e475e0f11951b3767336632971
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>