Allow deleting multiple branches at once
From time to time, one wants to cleanup old and obsolete branches.
To help the user with this task, it is now possible to select multiple local
branches and delete them all at once. To improve user experience, the
affected branches will be shown again in the dialog before confirmation.

Bug: 319246
Change-Id: I70ea9db24fafa90742ee147248558d8ffb011bc8
Signed-off-by: Benjamin Muskalla <bmuskalla@eclipsesource.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>