Don't start SourceBranchPage with a warning
Currently the SourceBranchPage starts with the warning that the
repository is empty as we're just about to fetch the refs. We should
only check if the repository is empty when we are done with fetching
its branches. This avoids a warning the user should only see in the case
the repository really is empty.

Change-Id: Ida4d4cb0ed0d1881d8a263d8b51caacbb5450fa7
Signed-off-by: Benjamin Muskalla <bmuskalla@eclipsesource.com>