Supportability: Provide a Trace Configuration Preference Page
Since we have the infrastructure in place, we can now switch debug
tracing on and off in a running IDE. This can be very helpful in
situations where we can't debug.

Unfortunately, there is no generic UI to do this, so this adds a
dialog to show the existing trace locations for the EGit UI and
EGit Core plug-ins and allows to activate them.

The dialog cannot be started from a "normal" UI control, but only
using CTRL-3 and triggering a dedicated command "Configure Git
Debug Trace" in order to avoid pollution of the Preference UI.

Change-Id: I4681acf85de8d05be6290d88acde5d99453ce89d
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>