[historyView] Refactoring second part
Use the HistoryPageInput as input in the GitHistoryPage. Calculate the
name only when the input changes. Add an adapter so that selection
in the history view automatically changes the input of the history view
if "link with selection" is active. Prepare HistoryPageInput to also allow
using Files instead of IResources.

Change-Id: I1b96ffc67cc7a2c66ff9a5bca445565483eaad27
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>