Optionally hard-wrap commit message
When "hard-wrap" is enabled in the new Commit Dialog preference page,
the commit message editor automatically inserts newlines as the user is
typing the commit message.

The text width is 70. If a line the user is typing would be longer than
that, a newline is automatically inserted before the last word.
Existing newlines are not touched, to allow for manual formatting (e.g.
ASCII graphics).

The editor shows the commit message in a configurable font that defaults
to the text font. This font is usually fixed-width, so that the margin
line can be shown after exactly 70 characters.

CQ: 4628
Bug: 329489
Change-Id: Ic8ae2f99b595f5f994af3cdb08b9363d2d53e685
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>