Support state [removed, untracked] in commit dialog
Files which are removed from the index but exist in the working
directory were displayed as "removed" in the commit dialog.
Committing such a change removed the affected files from the
repository. Now such files are displayed as "removed, untracked"
and on commit they are tracked as modification of the affected files.

Bug: 327087
Change-Id: I5e789ede05d9bf0487779ef023dbdc90a3197571
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>