Commit Dialog: reset author if amend is deselected
If in the commit dialag the amend option is
selected the author field is automatically updated
to the author of the previous commit. However if
the amend option got deselected again the author
of the previous commit stayed in the author field.
With this change the author field is now reset to
the original value if the amend option is
deselected.

Signed-off-by: Edwin Kempin <edwin.kempin@gmail.com>