Do not try to delete an already deleted folder
Originally, the remove command tries to delete first the working
directory (if repository has one) and then the metadata directory.
However the metadata directory could be contained by the working
directory and thus already deleted when the second deletion is called.
In this case an IOException will be thrown, to indicate that the
metadata directory could not be deleted. The correction simply checks
that the metadata directory exists before trying to delete it.

Bug: 332262
Change-Id: I6f49db1c8e730ab3c10d35e2af0e4dc81b5a223c
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>