Let user know when the configuration file is not editable
We may not be able to determine the location of the config file. This
would happen when no C Git can be found, since the "system wide"
configuration file is located relative to the git executable.

The config file may also be owned by root or be non-writeable
for other reasons.

Change-Id: I2384d83dea063ec2ffe11e54d77295f75edc8804
Signed-off-by: Robin Rosenberg <robin.rosenberg@dewire.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>