Create an internal extension point to extend the commit message
To integrate Mylyn's commit templates an extension point
is needed. During the preparation of the commit
dialogs the extensions using this extension points will be
requested for a message to add to the dialog. At the moment,
this extension point is meant for internal usage only.

Here is a sample configuration:
<extension
    point="org.eclipse.egit.ui.commitMessageProvider">
    <commitMessageProvider
        class="org.eclipse.egit.mylyn.CommitMessageProvider">
    </commitMessageProvider>
</extension>

Change-Id: Ib423ae5cbb8be1380744705c2fc57f52bec714f1
Signed-off-by: Benjamin Muskalla <benjamin.muskalla@tasktop.com>
Signed-off-by: Thorsten Kamann <thorsten@kamann.info>
Signed-off-by: Chris Aniszczyk <caniszczyk@gmail.com>