Introduce a tree compare
This improves the "Compare with -> Branch or Tag..." action to also
work in folders and projects in the workspace as well as enabling
"Compare with each other" action in the HistoryView to appear
independent of the current input (currently, this is only available if
the input of the HistoryView is a file).

When the user issues this action, a dialog will appear asking to
select a branch or tag; afterwards, the relevant part of the
repository will be scanned for changes and the result will be shown in
a specific "Git Tree Compare" view which shows two trees with added,
deleted, and changed files.

The user can then navigate around; by double-clicking on a file, he
can open a CompareEditor with the two respective version.

Bug: 304776
Bug: 315585
Change-Id: I95e44191e9d8cee5ce96fb84c7d7755b6b291dca
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>