[historyView] Performance first part
This improves the performance of the CommitMessageViewer by
only rendering the commit message and other meta information by
default. The diffs for the files in the commit are not rendered
immediately when clicking on a commit, but only when the user selects
files in the file list.

This results in dramatic performance improvements if there are commits
containing many files.

The user can still click "select all" in the file list to get all the
diffs into the view.

In addition, the diffs are rendered asynchronously letting the user
cancel the operation when it takes too long.

Some tracing was added to help maintenance in the future.

Bug: 337081
Change-Id: Idc2cf4119c8b34e048c3c345dedbd79c336b7063
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>