Enable cherry-pick conflict resolution
When a cherry-pick fails due to conflicts, the user has to resolve the
situation. With this change, the repository is set to the "Conflicts"
state and the merge tool is enabled.

Also, the original commit message and author is preset in the commit
dialog. The message includes a "Conflicts" section as in C git.

This depends on JGit change I947967fdc2f1d55016c95106b104c2afcc9797a1.

Bug: 339092
Change-Id: I5c5b98c98cf8deefa27de6cb94afa189a2cba218
Signed-off-by: Robin Stocker <robin@nibor.org>