Refactoring Branch operations (3): cleanup
Add a "Rename Branch" dialog and remove the BranchSelectionDialog. Use
an "Advanced" sub-menu in the Team menu to add actions for "Delete
Branch and "Rename Branch" and remove the old "Branch..." action from
the Team menu.

Move the "Tag..." action also under the "Advanced" sub-menu.

Change the "Branch..." action in the action set to use the new
"Checkout Branch" dialog.

In a subsequent change, we could also move the "Assume Unchanged/No
Assume Unchanged" actions to the "Advanced" sub-menu.

Change-Id: I6884d6d2f6870f0440a59c29e8416da882e2cf71
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Stefan Lay <stefan.lay@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>