Use <user_home>/git as default-default repository path
This is better than the currently used <user_home>, as this is very
likely to result in Repositories being created directly under
<user_home>. We should instead support creation of the Repositories
under their own folder.

The folder is created immediately during startup, otherwise the tools
that will use this folder won't be able to use it, e.g. point a
DirectoryDialog to this folder when the user hits "Browse..." etc. If
the folder can not be created during startup, we just log an error.

Change-Id: I2aee1b053cf991db064ece28231218bb4db399bb
Signed-off-by: Mathias Kinzler <mathias.kinzler@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>