Fix "Open in Commit Viewer" from resource-filtered History view
When looking at the history of a single resource in the History view and
then opening a commit in the Commit Viewer, the Files section would show
invalid changes.

The problem was that the parent of the RevCommit was set to the previous
commit for the same resource and therefore the diff included all changes
between those commits (not only to the previous one).

We fix this by re-parsing the commit with a clean RevWalk before showing
it.

Change-Id: I24dbad00575e4c6b38da72c7b6ea6b90b30c5b33
Signed-off-by: Robin Stocker <robin@nibor.org>
Signed-off-by: Kevin Sawicki <kevin@github.com>