[historyView] Reveal selected commit on filter change
When looking at the commit history of a resource, a common use case is
to select a commit and then wanting to see how this commits fits into
the history of the whole repository. With this change, this is now
easier because the selected commit is automatically revealed when
changing the filter.

Bug: 324736
Change-Id: I3ee99773af88cf073315c3f7bf048b0fe2dd694b
Signed-off-by: Robin Stocker <robin@nibor.org>