Run Test Suite AllNonSWTTests in Maven Build
The test suite AllNonSWTTests was not part of the build.
To execute AllNonSWTTests a new test suite AllSWTAndNonSWTTests was
created that combines the suites AllLocalTests and AllNonSWTTests.
The Eclipse launch configuration "All-Non-SWT-Tests" was changed to run
outside the UI thread. This was done because the suite AllNonSWTTests
does not require the UI thread and does not run stable in the UI thread.

Change-Id: I2b40859568ec2d0fa4efc8b3da5ba69ae49d3d73
Signed-off-by: Jens Baumgart <jens.baumgart@sap.com>
Signed-off-by: Matthias Sohn <matthias.sohn@sap.com>